+++ osidb/__init__.py
@@ -2,5 +2,5 @@
 osidb version
 """
 
-__version__ = "5.5.0"
+__version__ = "5.5.1"
 VERSION = __version__