+++ docs/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- cve.org collect for mitre_cve_editor (OSIDB-4786)
+
+
 ## [5.5.0] - 2026-02-06
 ### Fixed
 - Fetching specific events through the /audit endpoint now works (OSIDB-4777)