+++ docs/CHANGELOG.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Uniqueness on purl or ps_component for affects (OSIDB-4644)
 - Make `ps_module` read-only in the affect serializer (OSIDB-4554)
 - Allow filtering by PURL fields (OSIDB-4728)
+- Improve promote performance with prefetch (OSIDB-4678)
 
 ### Added
 - Add `in` filter for API views (OSIDB-4588)